+++ src/laconic-hover.js
@@ -1,6 +1,7 @@
 import tippy from "tippy.js";
 import { hoverTemplate } from "./lib/templates";
-import { Trope, darkModeEnabled } from "./lib/tvtropes-api";
+import { Trope } from "./lib/tvtropes-api";
+import { darkModeEnabled } from "./lib/utils";
 import Constants from "./lib/constants";
 import "tippy.js/dist/tippy.css";
 import "tippy.js/dist/themes/light.css";
